requirements.txt
Django==4.0
psycopg2_binary==2.9.2
djangorestframework==3.13.1
django-filter==21.1
djangorestframework-simplejwt==5.2.2